pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel>=0.37.0"]
build-backend = "setuptools.build_meta"


[project]
name = "pynest-api"
description = "PyNest is a FastAPI Abstraction for building microservices, influenced by NestJS."
readme = "README.md"
requires-python = ">=3.8.1"
license = { file = "LICENSE" }
authors = [
    { name = "Itay Dar", email = "itay2803@gmail.com" },
]
dynamic = ["version"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "click>=8.1.6",
    "fastapi==0.95.1",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.23.1",
    "PyYAML>=6.0.1",
    "astor>=0.8.1",
    "black>=23.11.0",
    "injector>=0.21.0",
    "pydantic<2.0.0",
    "sqlalchemy == 2.0.19",
    "alembic == 1.7.4",
]

[tool.setuptools.dynamic]
version = { attr = "nest.__init__.__version__" }

[tool.pip]
index-url = "https://pypi.org/simple"
trusted-host = ["pypi.org", "files.pythonhosted.org"]

[tools.black]
force-exclude = '''
/(
  | /*venv*
  | /.git
  | /dist
  | /pytest_compare.egg-info
  | *.bac
  | .mypy_cache
  | .coverage
  | /htmlcov
  | /docs
  | /site
)/
'''

[project.optional-dependencies]
test = [
    "pytest == 6.2.5",
]

orm = [
    "sqlalchemy == 2.0.19",
    "alembic == 1.7.4",
]
mongo = [
    "pymongo == 3.12.0",
    "motor == 3.2.0",
    "beanie == 1.20.0",
]

[project.scripts]
pynest = "nest.cli.cli:nest_cli"

[tool.setuptools.packages.find]
include = ["nest*"]
namespaces = false

[tool.mypy]
exclude = [
    "/*venv*"
]
ignore_missing_imports = true

[project.urls]
"Homepage" = "https://github.com/PythonNest/PyNest"
"Documentation" = "https://pythonnest.github.io/PyNest/"